+++ inst/include/rxode2parse_control.h
@@ -145,8 +145,9 @@
 #define RxMv_stateOrd 27
 #define RxMv_lhsOrd 28
 #define RxMv_splitBolus 29
-#define RxMv_timeId 30
-#define RxMv_md5 31
+#define RxMv_strCmpParams 30
+#define RxMv_timeId 31
+#define RxMv_md5 32
 #define RxMvFlag_ncmt 0
 #define RxMvFlag_ka 1
 #define RxMvFlag_linB 2