+++ src/parseLogical.h
@@ -40,8 +40,8 @@ static inline int isOperatorOrPrintingIdentifier(nodeInfo ni, const char *name){
 
 static inline int isSkipChild(nodeInfo ni, const char *name, int i) {
   return ((i == 3 || i == 4 || i < 2) &&
-	  (nodeHas(derivative) ||nodeHas(fbio) || nodeHas(alag) ||
-	   nodeHas(rate) || nodeHas(dur))) ||
+          (nodeHas(derivative) ||nodeHas(fbio) || nodeHas(alag) ||
+           nodeHas(rate) || nodeHas(dur))) ||
     ((i == 3 || i < 2) && nodeHas(der_rhs)) ||
     (nodeHas(dfdy)     && i< 2)  ||
     (nodeHas(dfdy_rhs) && i< 2) ||
@@ -124,23 +124,82 @@ static inline int handleIfElse(nodeInfo ni, char *name, int i) {
   return 0;
 }
 
+static inline void clearStringCmpCurrent(void) {
+  tb.strCmpCurCov = NULL;
+  tb.strCmpCurStr = NULL;
+  tb.strCmpCurType = -1;
+}
+
+static inline int isStringCmpId(const char *v) {
+  return !strcmp(v, "id") || !strcmp(v, "ID") || !strcmp(v, "Id");
+}
+
+static inline void normalizeStringCmpValue(char *v) {
+  size_t n = strlen(v);
+  if (n >= 2 && ((v[0] == '"' && v[n-1] == '"') ||
+                 (v[0] == '\'' && v[n-1] == '\''))) {
+    memmove(v, v+1, n-2);
+    v[n-2] = 0;
+  }
+}
+
+static inline int getStringCmpCovIndex(const char *cov) {
+  for (int i = 0; i < tb.strCmp.n; ++i) {
+    if (!strcmp(tb.strCmp.line[i], cov)) {
+      return i;
+    }
+  }
+  if (tb.strCmp.n + 1 > tb.allocSC) {
+    tb.allocSC += MXDER;
+    tb.strCmpN = R_Realloc(tb.strCmpN, tb.allocSC, int);
+  }
+  addLine(&(tb.strCmp), "%s", cov);
+  tb.strCmpN[tb.strCmp.n-1] = 0;
+  return tb.strCmp.n-1;
+}
+
+static inline int addStringCmpValue(const char *cov, char *val) {
+  int covIndex = getStringCmpCovIndex(cov);
+  int valueIndex = 0;
+  char *normVal = R_Calloc(strlen(val) + 1, char);
+  strcpy(normVal, val);
+  normalizeStringCmpValue(normVal);
+  for (int i = 0; i < tb.strCmpVal.n; ++i) {
+    if (tb.strCmpValI[i] == covIndex) {
+      valueIndex++;
+      if (!strcmp(tb.strCmpVal.line[i], normVal)) {
+        R_Free(normVal);
+        return valueIndex;
+      }
+    }
+  }
+  if (tb.strCmpVal.n + 1 > tb.allocSCV) {
+    tb.allocSCV += MXDER;
+    tb.strCmpValI = R_Realloc(tb.strCmpValI, tb.allocSCV, int);
+  }
+  tb.strCmpValI[tb.strCmpVal.n] = covIndex;
+  tb.strCmpN[covIndex] += 1;
+  addLine(&(tb.strCmpVal), "%s", normVal);
+  R_Free(normVal);
+  return valueIndex + 1;
+}
+
 static inline int handleStringEqualRhs(nodeInfo ni, char *name, int i, D_ParseNode *xpn) {
   if (nodeHas(equality_str1)){
     char *v = (char*)rc_dup_str(xpn->start_loc.s, xpn->end);
     switch(i) {
     case 0:
       // string
-      aAppendN("_cmp1(", 6);
-      sAppend(&sb, "%s, ", v);
-      sAppend(&sbDt, "%s, ", v);
+      clearStringCmpCurrent();
+      tb.strCmpCurStr = v;
       sAppend(&sbt, "%s", v);
       /* Free(v); */
       return 1;
     case 1:
       if (!strcmp(v, "==")) {
-	aAppendN("1, ", 3);
+        tb.strCmpCurType = 1;
       } else {
-	aAppendN("0, ", 3);
+        tb.strCmpCurType = 0;
       }
       sAppend(&sbt, "%s", v);
       /* Free(v); */
@@ -148,14 +207,26 @@ static inline int handleStringEqualRhs(nodeInfo ni, char *name, int i, D_ParseNo
     case 2:
       // identifier_r
       // val, valstr
-      if (!strcmp(v, "id") || !strcmp(v, "ID") || !strcmp(v, "Id")){
-	aAppendN("(&_solveData->subjects[_cSub])->idReal, \"ID\")", 45);
-	sAppendN(&sbt, "ID", 2);
+      tb.strCmpCurCov = v;
+      if (isStringCmpId(v)){
+        aAppendN("_cmp1(", 6);
+        sAppend(&sb, "%s, %d, (&_solveData->subjects[_cSub])->idReal, \"ID\")",
+                tb.strCmpCurStr, tb.strCmpCurType);
+        sAppend(&sbDt, "%s, %d, (&_solveData->subjects[_cSub])->idReal, \"ID\")",
+                tb.strCmpCurStr, tb.strCmpCurType);
+        sAppendN(&sbt, "ID", 2);
+        clearStringCmpCurrent();
       } else {
-	if (new_or_ith(v)) addSymbolStr(v);
-	sAppend(&sb, "%s, \"%s\")", v, v);
-	sAppend(&sbDt, "%s, \"%s\")", v, v);
-	sAppend(&sbt, "%s", v);
+        if (new_or_ith(v)) addSymbolStr(v);
+        int cmpInt = 0;
+        if (tb.strCmpCurStr != NULL) {
+          cmpInt = addStringCmpValue(tb.strCmpCurCov, (char*)tb.strCmpCurStr);
+        }
+        aAppendN("_cmp1d(", 7);
+        sAppend(&sb, "%s, %d, %s, %d)", tb.strCmpCurStr, tb.strCmpCurType, v, cmpInt);
+        sAppend(&sbDt, "%s, %d, %s, %d)", tb.strCmpCurStr, tb.strCmpCurType, v, cmpInt);
+        sAppend(&sbt, "%s", v);
+        clearStringCmpCurrent();
       }
       return 1;
     }
@@ -168,29 +239,44 @@ static inline int handleStringEqualLhs(nodeInfo ni, char *name, int i, D_ParseNo
     char *v = (char*)rc_dup_str(xpn->start_loc.s, xpn->end);
     switch(i) {
     case 0:
-      aAppendN("_cmp2(", 6);
-      if (!strcmp(v, "id") || !strcmp(v, "ID") || !strcmp(v, "Id")){
-        aAppendN("(&_solveData->subjects[_cSub])->idReal, \"ID\", ", 46);
+      clearStringCmpCurrent();
+      tb.strCmpCurCov = v;
+      if (isStringCmpId(v)){
         sAppendN(&sbt, "ID", 2);
       } else {
         if (new_or_ith(v)) addSymbolStr(v);
-        sAppend(&sb, "%s, \"%s\", ", v, v);
-        sAppend(&sbDt, "%s, \"%s\", ", v, v);
         sAppend(&sbt, "%s", v);
       }
       return 1;
     case 1:
       if (!strcmp(v, "==")) {
-        aAppendN("1, ", 3);
+        tb.strCmpCurType = 1;
       } else {
-        aAppendN("0, ", 3);
+        tb.strCmpCurType = 0;
       }
       sAppend(&sbt, "%s", v);
       return 1;
     case 2:
-      sAppend(&sb, "%s)", v);
-      sAppend(&sbDt, "%s)", v);
+      tb.strCmpCurStr = v;
+      if (isStringCmpId(tb.strCmpCurCov)) {
+        aAppendN("_cmp2(", 6);
+        sAppend(&sb, "(&_solveData->subjects[_cSub])->idReal, \"ID\", %d, %s)",
+                tb.strCmpCurType, v);
+        sAppend(&sbDt, "(&_solveData->subjects[_cSub])->idReal, \"ID\", %d, %s)",
+                tb.strCmpCurType, v);
+      } else {
+        int cmpInt = 0;
+        if (tb.strCmpCurCov != NULL) {
+          cmpInt = addStringCmpValue(tb.strCmpCurCov, v);
+        }
+        aAppendN("_cmp2d(", 7);
+        sAppend(&sb, "%s, \"%s\", %d, %d)",
+                tb.strCmpCurCov, tb.strCmpCurCov, tb.strCmpCurType, cmpInt);
+        sAppend(&sbDt, "%s, \"%s\", %d, %d)",
+                tb.strCmpCurCov, tb.strCmpCurCov, tb.strCmpCurType, cmpInt);
+      }
       sAppend(&sbt, "%s", v);
+      clearStringCmpCurrent();
       return 1;
     }
   }
@@ -212,8 +298,8 @@ static inline int assertLogicalNoWhileElse(nodeInfo ni, char *name, int i, D_Par
       v = (char*)rc_dup_str(xpn2->start_loc.s, xpn2->end);
       if (v[0] == 0) {
       } else {
-	updateSyntaxCol();
-	trans_syntax_error_report_fn(_("'while' cannot be followed by 'else' (did you mean 'if'/'else')"));
+        updateSyntaxCol();
+        trans_syntax_error_report_fn(_("'while' cannot be followed by 'else' (did you mean 'if'/'else')"));
       }
     }
     return 1;